frontend: prevent scroll when slash menu opens
