+++ frontend/packages/editor/src/block-utils.ts
@@ -1,0 +1,34 @@
+import {Block as BlockNoteBlock, BlockNoteEditor} from './blocknote'
+import {HMBlockSchema} from './schema'
+import {useEffect, useState} from 'react'
+import {getBlockInfoFromPos} from './blocknote'
+
+export function useSelected(
+  block: BlockNoteBlock<HMBlockSchema>,
+  editor: BlockNoteEditor<HMBlockSchema>,
+) {
+  const [selected, setSelected] = useState(false)
+  const tiptapEditor = editor._tiptapEditor
+  const selection = tiptapEditor.state.selection
+
+  useEffect(() => {
+    if (editor) {
+      const selectedNode = getBlockInfoFromPos(
+        tiptapEditor.state.doc,
+        tiptapEditor.state.selection.from,
+      )
+      if (selectedNode && selectedNode.id) {
+        if (
+          selectedNode.id === block.id &&
+          selectedNode.startPos === selection.$anchor.pos
+        ) {
+          setSelected(true)
+        } else if (selectedNode.id !== block.id) {
+          setSelected(false)
+        }
+      }
+    }
+  }, [selection])
+
+  return selected
+}
